+++ scripts/guide/main.ts
@@ -1,4 +1,4 @@
-import { Pages } from "@lpsmods/mc-utils";
+import { Icon, Pages } from "@lpsmods/mc-utils";
 import { blocks } from "./blocks";
 import { changelogs } from "./changelogs";
 import { items } from "./items";
@@ -7,7 +7,12 @@ export const pages: Pages = {
   home: {
     title: "guide.common.guide_book",
     body: "#desc",
-    buttons: ["blocks", "items", "changelogs"],
+    buttons: ["blocks", "items", "creator", "changelogs"],
+  },
+  creator: {
+    title: "#creator",
+    icon: Icon.Comment,
+    body: "#creator.desc",
   },
   ...blocks,
   ...items,