+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] (1.21.132) - 1/18/2026
+
+### General
+
+- Added `Creator` tab to the guide book.
+- Added public API to add your own carving knife recipes.
+
 ## [2.0.0] (1.21.101) - 10/24/2025
 
 > NOTE: This update is not compatible with older versions.